修改

+++ win_build.py
@@ -1,49 +1,32 @@
 import os
 import shutil
 import subprocess
-import time
+
+subprocess.run(r"node C:\Users\chiro\AppData\Local\Programs\Xmind\resources\app\main\build\build.js")
+
 
 os.environ["GOARCH"] = "amd64"
 os.environ["GOOS"] = "windows"
 
 xmindActiveSrc = r"C:\Users\chiro\GolandProjects\xmindActive\cmd\xmindActive"
+activeFileDist = os.path.join(xmindActiveSrc,"dist","xmindActive_Win.exe")
+
 xmindUpdateSrc = r"C:\Users\chiro\GolandProjects\xmindActive\cmd\xmindUpdate"
+updateFileDist = os.path.join(xmindActiveSrc,"asset","xmindUpdate.exe")
 
-# 读取 hook.js 和 init.js 的内容
-with open("./hook.js", "r", encoding="utf-8") as f:
-    hook_content = f.read()
-with open("./init.js", "r", encoding="utf-8") as f:
-    init_content = f.read()
-xmind_content = hook_content + init_content
-# 将拼接后的内容保存为 xmind.b.js
-with open("xmind.b.js", "w", encoding="utf-8") as f:
-    f.write(xmind_content)
-# 执行 javascript-obfuscator
-os.system("javascript-obfuscator xmind.b.js --config ob.json -o xmind.b.js")
-# 读取 js-yaml.min.js 的内容
-with open("./lib/js-yaml.min.js", "r", encoding="utf-8") as f:
-    yaml_content = f.read()
-# 读取 xmind.b.js 的内容
-with open("xmind.b.js", "r", encoding="utf-8") as f:
-    xmind_b_content = f.read()
-# 在 xmind.b.js 文件前面插入 js-yaml.min.js 的内容
-new_content = yaml_content + xmind_b_content
-# 将新内容保存到 xmind.b.js 文件中
-with open("./asset/xmind.b.js", "w", encoding="utf-8") as f:
-    f.write(new_content)
+hookFilePatch = r"C:\Users\chiro\GolandProjects\xmindActive\cmd\hookFilePatch"
 
-# 睡眠一秒
-time.sleep(1)
+
+
 
 # 切换到项目根目录
 os.chdir(r"C:\Users\chiro\GolandProjects\xmindActive")
 
 # 获取脚本运行当前路径
 script_path = os.path.dirname(os.path.abspath(__file__))
-source_dir = os.path.join(script_path, "win/asset")
+source_dir = os.path.join(script_path, "asset")
 target_dirs = [
-    os.path.join(xmindActiveSrc, "asset"),
-    os.path.join(xmindUpdateSrc, "asset")
+    os.path.join(hookFilePatch, "asset"),
 ]
 # 复制文件
 for target_dir in target_dirs:
@@ -55,17 +38,13 @@ for target_dir in target_dirs:
             shutil.copy(full_file_name, target_dir)
             print(f"Copied {full_file_name} to {target_dir}")
 
-updateFileSrc = r"C:\Users\chiro\GolandProjects\xmindActive\cmd\xmindUpdate"
-updateFileDist = (
-    r"C:\Users\chiro\GolandProjects\xmindActive\cmd\xmindActive\asset\xmindUpdate.exe"
-)
-activeFileSrc = r"C:\Users\chiro\GolandProjects\xmindActive\cmd\xmindActive"
-activeFileDist = r"C:\Users\chiro\GolandProjects\xmindActive\cmd\xmindActive\dist\xmind_server_changes_win.exe"
+
 outfile = os.path.join(
     os.environ["USERPROFILE"],
     "Desktop",
-    "xmind_server_changes_win.exe",
+    "xmindActive_Win.exe",
 )
+
 # 定义要执行的 garble 和 upx 命令
 commands = [
     [
@@ -74,7 +53,7 @@ commands = [
         "build",
         "-o",
         updateFileDist,
-        updateFileSrc,
+        xmindUpdateSrc,
     ],
     [
         "upx",
@@ -88,7 +67,7 @@ commands = [
         "build",
         "-o",
         activeFileDist,
-        activeFileSrc,
+        xmindActiveSrc,
     ],
     [
         "upx",
@@ -99,7 +78,8 @@ commands = [
         activeFileDist,
     ],
 ]
-os.remove(outfile)
+if os.path.exists(outfile):
+    os.remove(outfile)
 # 执行命令
 for i, command in enumerate(commands):
     try:
